Added weapon sprites, flute

+++ OutOfTune/Assets/Scripts/Items/Weapons/CymbalMine.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class CymbalMine : MonoBehaviour
 {
 
-    public int damage = 5;
+    public float damage = 5;
 
     void OnCollisionEnter2D(Collision2D collision)
     {
